mpi/examples: minor fixes

+++ mpi/examples/reduction/mpi_reduction_kernels.c
@@ -52,7 +52,7 @@ void dot_cpu_func(void *descr[], void *cl_arg)
 	long int *local_x = (long int *)STARPU_VECTOR_GET_PTR(descr[0]);
 	unsigned n = STARPU_VECTOR_GET_NX(descr[0]);
 
-	long int *dot = (int *)STARPU_VARIABLE_GET_PTR(descr[1]);
+	long int *dot = (long int *)STARPU_VARIABLE_GET_PTR(descr[1]);
 
 //	_DISPLAY("Before dot=%ld (adding %d elements...)\n", *dot, n);
 	unsigned i;
